requirements.txt
requests>=2.28.0      # Для HTTP-запросов к API
schedule>=1.2.0       # Для планирования задач
pyyaml>=6.0           # Для чтения YAML-конфигов
python-dotenv>=1.0.0 # Для работы с .env файлами
pandas>=1.5.0         # Для работы с таблицами
beautifulsoup4>=4.11.0 # Для парсинга HTML
lxml>=4.9.0           # Для быстрого парсинга
